Solution5/backend/requirements.txt
fastapi
uvicorn
openai
pyodbc
azure-search-documents
numpy==1.26.4
azurefunctions-extensions-http-fastapi
azure-functions
starlette
msal
requests
langchain
langgraph
azure.identity
azure-mgmt-resource
azure.keyvault.secrets
azure.mgmt.network
pydantic
typing-extensions
opencensus-ext-azure
opencensus
azure.storage.blob
azure.data.tables
dsnparse
llama-index
tiktoken